Updated to the latest version of BotBase

+++ CountingBot/CommandHandler.cs
@@ -6,6 +6,7 @@ using System;
 using System.Collections.Generic;
 using System.Reflection;
 using System.Threading.Tasks;
+using static CountingBot.DatabaseManager;
 
 namespace CountingBot
 {
@@ -14,31 +15,31 @@ namespace CountingBot
         public const string prefix = "\\";
         public static int argPos = 0;
 
-        private readonly DiscordSocketClient _client;
-        private readonly CommandService _commands;
-        private readonly IServiceProvider _services;
+        private readonly DiscordSocketClient client;
+        private readonly CommandService commands;
+        private readonly IServiceProvider services;
 
         public CommandHandler(DiscordSocketClient client, IServiceProvider services)
         {
-            _client = client;
-            _services = services;
+            this.client = client;
+            this.services = services;
 
             CommandServiceConfig config = new CommandServiceConfig()
             {
                 DefaultRunMode = RunMode.Async
             };
-            _commands = new CommandService(config);
+            commands = new CommandService(config);
         }
 
         public async Task InitCommandsAsync()
         {
-            _client.Connected += SendConnectMessage;
-            _client.Disconnected += SendDisconnectError;
-            _client.MessageReceived += HandleCommandAsync;
-            _client.MessageReceived += HandleCountingAsync;
+            client.Connected += SendConnectMessage;
+            client.Disconnected += SendDisconnectError;
+            client.MessageReceived += HandleCommandAsync;
+            client.MessageReceived += HandleCountingAsync;
 
-            await _commands.AddModulesAsync(Assembly.GetEntryAssembly(), _services);
-            _commands.CommandExecuted += SendErrorAsync;
+            await commands.AddModulesAsync(Assembly.GetEntryAssembly(), services);
+            commands.CommandExecuted += SendErrorAsync;
         }
 
         private async Task SendErrorAsync(Optional<CommandInfo> info, ICommandContext context, IResult result)
@@ -51,18 +52,12 @@ namespace CountingBot
 
         private async Task SendConnectMessage()
         {
-            if (Program.isConsole)
-            {
-                await Console.Out.WriteLineAsync($"{SecurityInfo.botName} is online");
-            }
+            await Console.Out.WriteLineAsync($"{SecurityInfo.botName} is online");
         }
 
         private async Task SendDisconnectError(Exception e)
         {
-            if (Program.isConsole)
-            {
-                await Console.Out.WriteLineAsync(e.Message);
-            }
+            await Console.Out.WriteLineAsync(e.Message);
         }
 
         private async Task HandleCountingAsync(SocketMessage m)
@@ -72,10 +67,10 @@ namespace CountingBot
                 return;
             }
 
-            if (channel.Id == (await SetChannel.GetCountingChannelAsync(channel.Guild)).Id)
+            if (channel.Id == (await countingDatabase.Channels.GetCountingChannelAsync(channel.Guild)).Id)
             {
-                Task<int> nextCount = SetChannel.GetCountAsync(channel.Guild).ContinueWith(x => x.Result + 1);
-                Task<int> lastUserNum = GetUserCount.GetLastUserNumAsync(user);
+                Task<int> nextCount = countingDatabase.Channels.GetCountAsync(channel.Guild).ContinueWith(x => x.Result + 1);
+                Task<int> lastUserNum = countingDatabase.UserCounts.GetLastUserNumAsync(user);
                 if (m.Content != (await nextCount).ToString() || await lastUserNum + 1 == await nextCount)
                 {
                     await msg.DeleteAsync();
@@ -83,29 +78,32 @@ namespace CountingBot
                 else
                 {
                     await Task.WhenAll(
-                        SetChannel.IncrementCountAsync(user.Guild),
-                        GetUserCount.IncrementUserCountAsync(user, await nextCount)
+                        countingDatabase.Channels.IncrementCountAsync(user.Guild),
+                        countingDatabase.UserCounts.IncrementUserCountAsync(user, await nextCount)
                     );
                 } 
             }
         }
 
+        private async Task<bool> CanBotRunCommandsAsync(SocketUserMessage msg) => await Task.Run(() => msg.Author.Id == client.CurrentUser.Id);
+        private async Task<bool> ShouldDeleteBotCommands() => await Task.Run(() => true);
+
         private async Task HandleCommandAsync(SocketMessage m)
         {
-            if (!(m is SocketUserMessage msg) || (msg.Author.IsBot && msg.Author.Id != _client.CurrentUser.Id))
+            if (!(m is SocketUserMessage msg) || (msg.Author.IsBot && !await CanBotRunCommandsAsync(msg)))
             {
                 return;
             }
 
-            SocketCommandContext Context = new SocketCommandContext(_client, msg);
-            bool isCommand = msg.HasMentionPrefix(_client.CurrentUser, ref argPos) || msg.HasStringPrefix(prefix, ref argPos);
+            SocketCommandContext Context = new SocketCommandContext(client, msg);
+            bool isCommand = msg.HasMentionPrefix(client.CurrentUser, ref argPos) || msg.HasStringPrefix(prefix, ref argPos);
 
             if (isCommand)
             {
-                var result = await _commands.ExecuteAsync(Context, argPos, _services);
+                var result = await commands.ExecuteAsync(Context, argPos, services);
 
                 List<Task> cmds = new List<Task>();
-                if (msg.Author.IsBot)
+                if (msg.Author.IsBot && await ShouldDeleteBotCommands())
                 {
                     cmds.Add(msg.DeleteAsync());
                 }
